Pacify cron jobs
The `startingDeadlineSeconds` and `concurrencyPolicy` attributes of
a `CronJob` specification are super-important for failure handling.

Modified the Deep Security Smart Check cron jobs to set these
attributes nicely so that in failure modes we don't end up with
thousands of pending pods.

Also modified the default frequency of the metrics collector job to
hourly instead of every minute.