Add support for disabling the IPv6 listener in the proxy pod
We have received reports that deploying Deep Security Smart Check
on some Kubernetes on-premises installations fails due to IPv6 not
being enabled on the worker nodes.

This change adds the ability to disable the IPv6 listeners in
the proxy pod.

This does not change any externally-visible behaviour, most notably
it will not affect whether the load balancer created by Kubernetes
supports IPv6 externally.