Simplify certificate replacement procedure
The certificate replacement procedure directs users to use
`kubectl create secret tls` to create a secret with the certificate
that the `proxy` and `registry` services will use.

Before, the procedure required users to provide not only the
name of the secret, but also the names of the items within the
secret that corresponded to the certificate and private key.

Because `kubectl create secret tls` always uses the same names
for the certificate and private key items witin the secret, this
change sets the defaults in the template to those names.

Users who create the secret differently can still use the
`certificate.secret.certificate` and `certificate.secret.privateKey`
overrides to set the names of the items if they need to.