Trigger pod restart on proxy configuration changes
Kubernetes does not seem to automatically restart pods that have
environment variables which take their values from config maps
or secrets when the config map / secret is changed.

Adding a checksum environment variable for the outbound proxy
configuration so that everything that consumes the outbound proxy
environment variables will be restarted when the user changes
the configuration through a `helm upgrade`.