Trigger pod restart on database credential changes
Kubernetes does not seem to automatically restart pods that have
environment variables which take their values from config maps
or secrets when the config map / secret is changed.

Adding a checksum environment variable for the database credential
configuration so that everything that consumes the database credential
environment variables will be restarted when the user changes
the configuration through a `helm upgrade`.